guardar total bultos y total kgr en tabla order

+++ database/migrations/2024_09_08_023046_create_orders_table.php
@@ -25,7 +25,7 @@ return new class extends Migration
             $table->string('observacion')->nullable();
             $table->integer('estado')->default(0);
             $table->integer('totaBultos')->nullable();
-            $table->integer('totalKgr')->nullable();
+            $table->decimal('totalKgr')->nullable();
 
             // Relaciones con las direcciones de remitente y destinatario
             $table->foreign('remitente_direccion_id')
